Revert "LPS-99657 'test.internal' -> 'internal.test'"
This reverts commit a2b72aa54326a241c86f4d3e5e1b0687c66a99d8.
