+++ modules/apps/oauth2-provider/oauth2-provider-test/src/main/java/com/liferay/oauth2/provider/test/internal/TestApplication.java
@@ -12,33 +12,33 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.internal.test;
+package com.liferay.oauth2.provider.test.internal;
 
 import java.util.Collections;
 import java.util.Set;
 
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.core.Application;
 
 /**
- * @author Víctor Galán
+ * @author Carlos Sierra Andrés
  */
-public class TestPreviewURLApplication extends Application {
-
-	public TestPreviewURLApplication(String previewURL) {
-		_previewURL = previewURL;
-	}
-
-	@GET
-	public String getPreviewURL() {
-		return _previewURL;
-	}
+public class TestApplication extends Application {
 
 	@Override
 	public Set<Object> getSingletons() {
 		return Collections.<Object>singleton(this);
 	}
 
-	private final String _previewURL;
+	@GET
+	public String getString() {
+		return "get";
+	}
+
+	@POST
+	public String post(String post) {
+		return "post";
+	}
 
 }
