+++ modules/apps/oauth2-provider/oauth2-provider-test/src/main/java/com/liferay/oauth2/provider/test/internal/TestHeadHandlingApplication.java
@@ -12,25 +12,32 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.internal.test;
+package com.liferay.oauth2.provider.test.internal;
 
-import com.liferay.oauth2.provider.scope.RequiresNoScope;
-import com.liferay.oauth2.provider.scope.RequiresScope;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+import com.liferay.oauth2.provider.test.internal.annotation.CUSTOM;
 
 import java.util.Collections;
 import java.util.Set;
 
 import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import javax.ws.rs.HEAD;
+import javax.ws.rs.POST;
 import javax.ws.rs.core.Application;
 
 /**
  * @author Carlos Sierra Andrés
  */
-public class TestAnnotatedApplication extends Application {
+public class TestHeadHandlingApplication extends Application {
+
+	@CUSTOM
+	public String getCustomString() {
+		return "custom";
+	}
+
+	@HEAD
+	public String getHeadString() {
+		return "head";
+	}
 
 	@Override
 	public Set<Object> getSingletons() {
@@ -38,21 +45,13 @@ public class TestAnnotatedApplication extends Application {
 	}
 
 	@GET
-	@RequestScopeRead
 	public String getString() {
-		return "everything.read";
+		return "get";
 	}
 
-	@GET
-	@Path("/no-scope")
-	@RequiresNoScope
-	public String getStringNoScope() {
-		return "no-scope";
-	}
-
-	@RequiresScope("everything.read")
-	@Retention(RetentionPolicy.RUNTIME)
-	public static @interface RequestScopeRead {
+	@POST
+	public String post(String post) {
+		return "post";
 	}
 
 }
