+++ modules/apps/oauth2-provider/oauth2-provider-test/src/main/java/com/liferay/oauth2/provider/test/internal/annotation/CUSTOM.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.internal.test.annotation;
+package com.liferay.oauth2.provider.test.internal.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
